增加模版支持

+++ admin/handler/api.go
@@ -22,7 +22,7 @@ func Api(ctx *gin.Context) {
 	}
 
 	if count > 0 {
-		if err := query.Limit(req.PageSize).Offset(req.PageSize * (req.Page - 1)).Find(&list).Error; err != nil {
+		if err := query.Limit(int(req.PageSize)).Offset(int(req.PageSize * (req.Page - 1))).Find(&list).Error; err != nil {
 			response.FailWithError(ctx, err)
 			return
 		}
