+++ admin/handler/file.go
@@ -60,7 +60,7 @@ func FileList(ctx *gin.Context) {
 	}
 
 	if count > 0 {
-		if err := query.Order("id desc").Limit(req.PageSize).Offset(req.PageSize * (req.Page - 1)).Find(&list).Error; err != nil {
+		if err := query.Order("id desc").Limit(int(req.PageSize)).Offset(int(req.PageSize * (req.Page - 1))).Find(&list).Error; err != nil {
 			response.FailWithError(ctx, err)
 			return
 		}
